// src/components/TableComponents/StandardTableTypes.d.ts
export type GroupColorType = 'red' | 'gray'
export type groupAccordionType = {
  /** To add a `collapsed` class to expand/collapse the current clicked group's subrows */
  isCollapsed: boolean
  /** Unique key to identify each group */
  groupKey: string
}
type StickyTableConfigType = {
  /** The default for this is 1. If you need more than 1 sticky header, then use this prop. */
  header?: number
  /** The default for this is 1. If you need more than 1 sticky left column, then use this prop. */
  left?: number
  /** This should be set to 1 or more if the far right column has an action button in it. */
  right?: number
}
export type ConfigItemType<DataItem, StyleGeneric> = {
  /** Unique name for the column. Should be associated with the name that you are sorting on for the column. */
  name: SortColumnProps['propName']
  /** Unique label for the column headers. */
  label: SortColumnProps['label']
  /** Optionally add custom JSX. This will be rendered to the right of the column header. */
  columnHeaderSubContent?: React.ReactNode
  /** Used if you have multiple options to sort on within a column. */
  options?: SortColumnProps['options']
  /** The JSX that will be rendered in a cell. */
  cell: {
    className?: string | ((data: DataItem) => string)
    children: (data: DataItem, index?: number) => React.ReactNode
    /** Optional style overrides for a cell. */
    style?: StyleGeneric
  }
  /** Determines the static first column in mobile and also the column that will not be able to be removed in column customizations. */
  mainColumn?: boolean
  /** Used if the API requires a lowercase parameter to be passed in. */
  lowerCaseParam?: SortColumnProps['lowerCaseParam']
  /** Define the tooltip for a column. */
  tooltip?: SortColumnProps['tooltip']
  /** Class name for a column. */
  className?: string
  /** Prevent sorting for a column. */
  noSort?: boolean
  /** Column that has a single button. Used for display purposes in mobile. */
  isButton?: boolean
  /** Filter for a column. */
  filter?: SortColumnProps['filter']
  /** Style overrides for a column. */
  style?: StyleGeneric
}
export type StandardTableProps<
  DataItem,
  ConfigItem extends ConfigItemType<DataItem, Record<string, unknown>>
> = {
  /** Array of data passed into the Table */
  data: DataItem[]
  /** Table configuration that will handle column headers and rendering column children. */
  config: ConfigItem[]
  /** Unique key from our data objects. Used for `key` in our `map` functions. */
  dataKey: keyof DataItem
// ...
  totalRowKey?: keyof DataItem
// ...
}
// ...
export type GroupHeader<DataItem> = DataItem & {
  /** The string value to display in the group Header row (groupHeader) */
  groupHeader?: string | ((count: number) => string)
  /** Custom JSX to display info icon in the group Header row (groupHeader) */
  tooltipContent?: TooltipProps['tooltipContent']
  /** The color of the group header row */
  type?: GroupColorType
  /** boolean to add Clear button in the selection */
  includeClearButton?: boolean
  /** Function to perform the required action after clearing the selections */
  clearCallout?: () => void
  /** indicates if the item has been selected by the user using a checkbox */
  checked?: boolean
  /** Custom JSX at the last column of the group header to replace clearButton */
  customClearButton?: React.ReactNode
  /** Boolean value; only added to Header rows  */
  isGroupHeader?: boolean
  /** Toggle to display header details if header details exist */
  displayHeaderData?: boolean
  /** Custom JSX to display in the last column of the table */
  lastColumn?: React.ReactNode
  /**
   * Number to associate the group header with the right group; groups are identified in order
   * of appearance in the table (index + 1, so the first group is 1, the second is 2, etc.)
   */
  groupNum?: number
  /** Adds an expand/collapse functionality to entire groups  */
  groupAccordion?: groupAccordionType
}